# Conflicts:
#	i21Database/dbo/Stored Procedures/uspARInvoiceAgriCare.sql

+++ i21Database/dbo/Tables/tblIPContractHeaderStage.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE tblIPContractHeaderStage (
 	intContractHeaderStageId INT IDENTITY(1, 1)
-	,intDocNo INT
+	,intDocNo BIGINT
 	,strSender NVARCHAR(50) COLLATE Latin1_General_CI_AS
 	,strContractNo NVARCHAR(50) COLLATE Latin1_General_CI_AS
 	,dtmContractDate DATETIME
